Add PluginInfo.json and refactor project structure
