--- cmake/MSVCSettings.cmake
@@ -1,18 +1,0 @@
-if(MSVC)
-    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /DWIN32 /EHsc /MP /bigobj /permissive- /Zc:__cplusplus")
-	set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} /MDd")
-    set(CMAKE_CXX_FLAGS_RELWITHDEBINFO "${CMAKE_CXX_FLAGS_RELWITHDEBINFO} /MD")
-    set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} /MD")
-	set(CMAKE_EXE_LINKER_FLAGS_RELWITHDEBINFO "${CMAKE_EXE_LINKER_FLAGS_RELWITHDEBINFO} /DEBUG")
-	set(CMAKE_CXX_FLAGS_RELEASE_INIT "/GL")
-	set(CMAKE_SHARED_LINKER_FLAGS_INIT "/LTCG:PGOptimize")
-	
-	cmake_host_system_information(RESULT TOTAL_PHYSICAL_MEMORY QUERY TOTAL_PHYSICAL_MEMORY)
-	#message(STATUS "Available Memory = "${TOTAL_PHYSICAL_MEMORY} )
-	if(${TOTAL_PHYSICAL_MEMORY} LESS 16385)
-		set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} /MD /Ob0")
-		message(STATUS "Disabling Inline Function Expansion")
-	else()
-		set(CMAKE_CXX_FLAGS_RELEASE "${CMAKE_CXX_FLAGS_RELEASE} /MD")
-	endif()
-endif(MSVC)
