+++ src/H510XLoader/.editorconfig
@@ -1,0 +1,10 @@
+# To learn more about .editorconfig see https://aka.ms/editorconfigdocs
+root = true
+
+# All files
+[*]
+indent_style = tab
+
+# Cpp files
+[*.{h,cpp}]
+indent_size = 4
