+++ src/H510XLoader/PluginInfo.json
@@ -1,6 +1,10 @@
 {
     "name" : "H5 Loader (10X, Cell Ranger)",
     "menuName" : "H5 (.h5)",
-    "version" : "1.0",
+    "version" :  {
+        "plugin" : "1.0.0",
+        "core" : ["1.3"]
+    },
+    "type" : "Loader",
     "dependencies" : ["Points"]
 }
