--- src/H5ADLoader.json
@@ -1,6 +1,0 @@
-{
-    "name" : "H5AD Loader",
-    "menuName" : "H5AD (.h5ad)",
-    "version" : "1.0",
-    "dependencies" : ["Points"]
-}
