+++ src/H5ADLoader/PluginInfo.json
@@ -1,0 +1,10 @@
+{
+    "name" : "H5AD Loader",
+    "menuName" : "H5AD (.h5ad)",
+    "version" :  {
+        "plugin" : "1.0.0",
+        "core" : ["1.3"]
+    },
+    "type" : "Loader",
+    "dependencies" : ["Points"]
+}
