--- src/TOMELoader.json
@@ -1,6 +1,0 @@
-{
-    "name" : "TOME Loader",
-    "menuName" : "TOME (.tome)",
-    "version" : "1.0",
-    "dependencies" : ["Points"]
-}
