+++ src/TOMELoader/PluginInfo.json
@@ -1,0 +1,10 @@
+{
+    "name" : "TOME Loader",
+    "menuName" : "TOME (.tome)",
+    "version" :  {
+        "plugin" : "1.0.0",
+        "core" : ["1.3"]
+    },
+    "type" : "Loader",
+    "dependencies" : ["Points"]
+}
