+++ src/TOMELoader/TOMELoaderFactory.h
@@ -8,16 +8,16 @@ using namespace mv::plugin;
 // Factory
 // =============================================================================
 
-class H5ADLoaderFactory : public LoaderPluginFactory
+class TOMELoaderFactory : public LoaderPluginFactory
 {
 	Q_INTERFACES(mv::plugin::LoaderPluginFactory mv::plugin::PluginFactory)
 	Q_OBJECT
-	Q_PLUGIN_METADATA(	IID   "nl.lumc.H5ADLoaderFactory"
-						FILE  "H5ADLoader.json")
+	Q_PLUGIN_METADATA(	IID   "nl.lumc.TOMELoader"
+						FILE  "PluginInfo.json")
 
 public:
-	H5ADLoaderFactory(void) {}
-	~H5ADLoaderFactory(void) override {}
+	TOMELoaderFactory(void) {}
+	~TOMELoaderFactory(void) override {}
 
 	LoaderPlugin* produce() override;
